services/api/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
sentence-transformers
numpy
prometheus-client
azure-search-documents
